Change shape of some redux thunks' props
This is caused by Trello and the fact that it's either really hard, or, more likely, outright impossible to test it in the way it was until now. At least for me.

So now I decided to once again make this project a little more shitty and came up with this solution